pyproject.toml
[project]
name = "qiulao"
version = "0.1.0"
description = "AI-powered resume optimization system that helps users optimize their resumes based on job descriptions"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=24.1.0",
    "fastapi>=0.116.0",
    "jieba>=0.42.1",
    "jinja2>=3.1.6",
    "markdown>=3.8.2",
    "openai>=1.93.3",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "uvicorn[standard]>=0.35.0",
]
